Add scripts to package.json and tweak mr script

+++ scripts/deleteMaterialReports/index.js
@@ -29,45 +29,42 @@ const deleteMR = async () => {
 	jwt = await capi.login(process.env.CAPI_USER, process.env.CAPI_PASS);
 
 	// Grab material reports
-	logger.info('Fetching material reports from Canonn API');
+	logger.info('Fetching material reports older than ' + length + ' month');
 
 	// Grab total count of material reports
 	let mrCount = await capi.countMaterialReport(length);
 
 	logger.info(`There are ${mrCount} material reports to be deleted`);
 
-	logger.start('----------------');
-	logger.start(`Fetching first ${limit} material reports`);
-	logger.start('----------------');
-	while (keepGoing === true) {
-		let response = await capi.getMaterialReports(length, start);
-
-		for (i = 0; i < response.length; i++) {
-			logger.info(`Deleting Material Report ID: ${response[i].id} [${deleteCount + 1}/${mrCount}]`);
-			logger.info('--> Asking CAPI to delete');
-			let responseDelete = await capi.deleteMaterialReport(response[i].id, jwt);
-
-			if (responseDelete) {
-				logger.info('<-- Material Report deleted');
-			} else {
-				logger.warn('<-- Failed to delete Material Report');
+	if(mrCount > 0) {
+		logger.start('----------------');
+		logger.start(`Fetching first ${limit} material reports`);
+		logger.start('----------------');
+		while (keepGoing === true) {
+			let response = await capi.getMaterialReports(length, start);
+			logger.info(`Deleting ${response.length} Material Reports`)
+	
+			for (i = 0; i < response.length; i++) {
+				let responseDelete = await capi.deleteMaterialReport(response[i].id, jwt);
+	
+				if (!responseDelete) {
+					logger.warn('Failed to delete Material Report ID: ' + response[i].id)
+				}
+				deleteCount = (deleteCount + 1);
 			}
-			deleteCount = deleteCount + 1;
-			await delay(settings.global.delay);
-		}
-
-		if (response.length < limit) {
-			keepGoing = false;
-			logger.info('Deleted ' + deleteCount + ' material reports from the Canonn API');
-		} else {
-			logger.start('----------------');
-			logger.start(`Fetching next ${limit} material reports`);
-			logger.start('----------------');
-			start = start + limit;
-			await delay(settings.global.delay * 3);
+	
+			if (response.length < limit) {
+				keepGoing = false;
+				logger.info('Deleted ' + deleteCount + ' material reports from the Canonn API');
+			} else {
+				logger.start('----------------');
+				logger.start(`Fetching next ${limit} material reports`);
+				logger.start('----------------');
+				start = start + limit;
+				await delay(settings.global.delay * 3);
+			}
 		}
 	}
-
 	logger.stop('----------------');
 	logger.stop('Script Complete!');
 	logger.stop('----------------');
@@ -81,7 +78,13 @@ if (params.includes('--now'.toLowerCase()) === true) {
 
 if (isCron === true) {
 	logger.start('Starting in cron mode');
-	cron.schedule(settings.scripts[scriptName].cron[process.env.NODEID], () => {
+	let nodeID = 0
+
+	if (process.env.NODEID) {
+		nodeID = process.env.NODEID
+	}
+
+	cron.schedule(settings.scripts[scriptName].cron[nodeID], () => {
 		deleteMR();
 	});
 }
